+++ public/app.js
@@ -929,9 +929,14 @@ function resizeView() {
 let net = null;
 exports.net = net;
 
-let activateNet = () => {
-  if (typeof _manifest.default.net !== 'undefinde' && _manifest.default.net === true) {
-    var t = new _clientConfig.default();
+let activateNet = customConfig => {
+  if (typeof _manifest.default.net !== 'undefined' && _manifest.default.net === true) {
+    if (typeof customConfig !== 'undefined') {
+      var t = new customConfig();
+    } else {
+      var t = new _clientConfig.default();
+    }
+
     exports.net = net = new _net.Broadcaster(t);
     _manifest.default.network = net;
     console.info('Networking is active.', net);
@@ -5429,7 +5434,7 @@ class Position {
     this.targetX = newx;
     this.inMove = false;
 
-    if (_engine.net && _engine.net.connection && typeof em === 'undefined') {
+    if (_engine.net && _engine.net.connection && typeof em === 'undefined' && _manifest.default.scene[this.nameUniq].net.enable == true) {
       _engine.net.connection.send({
         netPos: {
           x: this.x,
@@ -5445,7 +5450,7 @@ class Position {
     this.y = newy;
     this.targetY = newy;
     this.inMove = false;
-    if (_engine.net && _engine.net.connection && typeof em === 'undefined') _engine.net.connection.send({
+    if (_engine.net && _engine.net.connection && typeof em === 'undefined' && _manifest.default.scene[this.nameUniq].net.enable == true) _engine.net.connection.send({
       netPos: {
         x: this.x,
         y: this.y,
@@ -5459,7 +5464,7 @@ class Position {
     this.z = newz;
     this.targetZ = newz;
     this.inMove = false;
-    if (_engine.net && _engine.net.connection && typeof em === 'undefined') _engine.net.connection.send({
+    if (_engine.net && _engine.net.connection && typeof em === 'undefined' && _manifest.default.scene[this.nameUniq].net.enable == true) _engine.net.connection.send({
       netPos: {
         x: this.x,
         y: this.y,
@@ -5477,7 +5482,7 @@ class Position {
     this.targetY = newy;
     this.targetZ = newz;
     this.inMove = false;
-    if (_engine.net && _engine.net.connection && typeof em === 'undefined') _engine.net.connection.send({
+    if (_engine.net && _engine.net.connection && typeof em === 'undefined' && _manifest.default.scene[this.nameUniq].net.enable == true) _engine.net.connection.send({
       netPos: {
         x: this.x,
         y: this.y,